// src/services/api/reportApi.ts
export interface TollReportPeriodCard {
  debitAmount?: number;
  creditAmount?: number;
  noOfTolls?: number;
  tollExpenses?: number;
  claimAmount?: number;
}